Establish a single entry point Scrutinizer and Composer Installation
